Refactored functions from utils, added move constructors for Cell and Region classes

+++ src/cpp/string.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 
 namespace cpp {
-    const std::string repeat(const std::string& string, int repeats);
+    const std::string repeat(const std::string& string, uint_fast8_t repeats);
 }
